// src/types.d.ts
export type Client = {
    id: string
    address: string
    businessName: string
    city: string
    contact: string
    email: string
    nit: string
    phone: string
    headquarters: Headquarter[]
    locations: Location[]
    equipments: Equipment[]
    user_app: UserApp
    fecha:string;
}

export type ClientResponse = {
    clients: Client[]
    fecha:string;
}

export type Equipment = {
    id: string
    name: string
    description: string
    serial: string
    image: string
    model: string
    type: string
    brand: string
    locationId: string
    location: {
        id: string
        locationName: string
    }
    headquarter: {
        id: string
        headName: string
    }
    client: {
        id: string
        businessName: string
        user_app: UserApp[]
    }
    createdAt:string;
}

export type Headquarter = {
    id: string
    address: string
    headName: string
    email: string
    phone: string
    isPrincipal: boolean
    clientId: string
    status: string
    Client: ClientRes
    locations: Location[]
    createdAt:string;
}

export type Location = {
    id: string
    registro: number
    locationName: string
    headquarterId: string
    description: string
    headquarter: HeadquarterRes
    client: ClientRes
    equipments: Equipment[]
    createdAt:string;
    // headquarter?: Omit<Headquarter, "isPrincipal" | "address" | "email" | "phone">
    // client?: Omit<Client, "nit" | "address" | "email" | "phone" | "city" | "contact"> & UserApp
}

export type UserApp = {
    user_id: string
    role_id: string
    role_name: string
    createdAt:string;
}
// ...
export interface HeadquarterRes {
    headName: string;
    address: string;
    email: string;
    phone: string;
    isPrincipal: boolean;
    clientId: string;
    Client: ClientRes;
}

export interface ClientRes {
    businessName: string;
    nit: string;
    address: string;
    email: string;
    phone: string;
    city: string;
    contact: string;
    user_app: UserAppRes[];
}

export interface UserAppRes {
    user_id: string;
    role_id: string;
    role_name: string;
}

export interface TechRes {
    techId: string;
    techName: string;
    techNumId: string;
}


export interface HeadRes {
    techId:    string;
    techName:  string;
    techNumId: string;
}

export interface EquiomentPayload {
    id?:             string
    name:        string;
    description: string;
    serial:      string;
    image?:       string;
    model:       string;
    type:        string;
    brand:       string;
    locationId:  string;

}